requirements-dev.txt
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
astunparse==1.6.3
azure-cognitiveservices-vision-face==0.5.0
azure-common==1.1.27
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.4
django-environ==0.8.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.2
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.5.0
isodate==0.6.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
msrest==0.6.21
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.3
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.4.1
